check 8 byte aligned

+++ hw5/ziting/check_8_byte_align.c
@@ -25,7 +25,7 @@ int main() {
   for (int i = 0; i < NUM_ALLOC; i++) {
     ptrs[i] = Mem_Alloc(BYTE);
     assert(ptrs[i] != NULL);
-    assert(ptrs[i] % 8 == 0);
+    assert(((long)ptrs[i]) % 8 == 0);
 
     Mem_Dump();
   }
